# lz4 decompress in file worker