@DmitryRomanov 2025-12-29
# Benches

## before (file.d v0.73.1)

```
go run ./bench/server
info	filebeat 50
info	started, waiting for first request
info	stats bytes=301Mb, requests=405, throughput: 302.67Mb/s
info	stats bytes=503Mb, requests=675, throughput: 252.73Mb/s
info	stats bytes=728Mb, requests=981, throughput: 242.69Mb/s
info	stats bytes=1044Mb, requests=1409, throughput: 260.88Mb/s
info	stats bytes=1309Mb, requests=1765, throughput: 261.98Mb/s
info	stats bytes=1605Mb, requests=2162, throughput: 267.71Mb/s
info	stats bytes=1872Mb, requests=2521, throughput: 267.31Mb/s
info	stats bytes=2186Mb, requests=2942, throughput: 273.08Mb/s
info	stats bytes=2434Mb, requests=3280, throughput: 270.53Mb/s
info	final result: 266.30Mb/s
```

## after
```
go run ./bench/server
info	filebeat 50
info	started, waiting for first request
info	stats bytes=276Mb, requests=369, throughput: 289.29Mb/s
info	stats bytes=518Mb, requests=695, throughput: 259.48Mb/s
info	stats bytes=809Mb, requests=1091, throughput: 270.03Mb/s
info	stats bytes=1028Mb, requests=1390, throughput: 257.19Mb/s
info	stats bytes=1327Mb, requests=1787, throughput: 266.75Mb/s
info	stats bytes=1593Mb, requests=2146, throughput: 265.78Mb/s
info	stats bytes=1920Mb, requests=2585, throughput: 274.27Mb/s
info	stats bytes=2244Mb, requests=3018, throughput: 280.38Mb/s
info	stats bytes=2568Mb, requests=3462, throughput: 285.32Mb/s
info	final result: 287.85Mb/s
```

```
goos: linux
goarch: amd64
pkg: github.com/ozontech/file.d/plugin/input/file
cpu: AMD EPYC 7763 64-Core Processor                
BenchmarkLightJsonReadPar-4   	       3	2166702207 ns/op	 669.79 MB/s	192067794 B/op	 2122678 allocs/op
PASS
ok  	github.com/ozontech/file.d/plugin/input/file	11.546s
```